Added support for changing area/room in Add/Edit entry form. Added check to ensure that at least one room is selected as edit_entry_handler.php does not catch this. Added wording to lang.en to cover the above check.
The first fix requires that javascript is enabled.  If it is not then you get the same form as you get currently.  If it is enabled and you have more than one area then you get an area select box which controls the rooms displayed in the room select box.